chores(mobile): Moves menus things back into components root
